backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
aiosqlite>=0.20.0
google-adk>=0.4.0
google-genai>=0.8.0
pydantic>=2.0.0
python-multipart>=0.0.9
